+++ hw12-webserver/src/main/java/ru/otus/hw/dao/model/User.java
@@ -2,7 +2,7 @@ package ru.otus.hw.dao.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ru.otus.hw.PasswordUtils;
+import ru.otus.hw.helpers.PasswordHelpers;
 import ru.otus.hw.orm.mapper.annotations.Id;
 
 /** @author sergey created on 03.02.19. */
@@ -17,13 +17,13 @@ public class User {
   private String passwordHash;
 
   public User(String login, String password) {
-    var salt = PasswordUtils.generateSalt();
+    var salt = PasswordHelpers.generateSalt();
     this.login = login;
     this.passwordSalt = salt;
-    this.passwordHash = PasswordUtils.hashPassword(salt, password);
+    this.passwordHash = PasswordHelpers.hashPassword(salt, password);
   }
 
   public boolean verifyPassword(String password) {
-    return PasswordUtils.hashPassword(passwordSalt, password).equals(passwordHash);
+    return PasswordHelpers.hashPassword(passwordSalt, password).equals(passwordHash);
   }
 }
