+++ hw12-webserver/src/main/java/ru/otus/hw/webserver/services/UserDaoProvider.java
@@ -1,0 +1,7 @@
+package ru.otus.hw.webserver.services;
+
+import ru.otus.hw.dao.UserDao;
+
+public interface UserDaoProvider {
+  UserDao getUserDao();
+}
